+++ ingredient_parser/__init__.py
@@ -1,8 +1,9 @@
-from ._common import SUPPORTED_LANGUAGES, show_model_card
+from ._common import SUPPORTED_LANGUAGES, UREG, show_model_card
 from .parsers import inspect_parser, parse_ingredient, parse_multiple_ingredients
 
 __all__ = [
     "SUPPORTED_LANGUAGES",
+    "UREG",
     "inspect_parser",
     "parse_ingredient",
     "parse_multiple_ingredients",