+++ tests/postprocess/test_fallback_pattern.py
@@ -72,7 +72,7 @@ class TestPostProcessor_fallback_pattern:
         """
         Test that imperial units are returned for 'cup'
         """
-        p = PostProcessor("", [], [], [], [], imperial_units=True)
+        p = PostProcessor("", [], [], [], [], volumetric_units_system="imperial")
         tokens = ["About", "2", "cup", "flour"]
         labels = ["COMMENT", "QTY", "UNIT", "B_NAME_TOK"]
         scores = [0.0] * len(tokens)
@@ -86,7 +86,7 @@ class TestPostProcessor_fallback_pattern:
                 confidence=0,
                 starting_index=1,
                 APPROXIMATE=True,
-                imperial_units=True,
+                volumetric_units_system="imperial",
             )
         ]